// src/app/model/user.ts
/* eslint-disable @typescript-eslint/naming-convention */
export class User {

    id: number;
    name: string;
    last_name: string;
    address: string;
    departament: string;
    neighborhood: string;
    country: string;
    email: string;
    phone_number: string;
    user_account_id: string;
    city: string;
    geo_location: string;

}